:truck: Rename: RecruitmentPageNameChange
